Refactored to use pre-existing TracerouteHop objects to store hop information
Fixed error in how parser stores hop information before saving
