+++ datastore/importers/nmdb-import-traceroute/nmdb-import-traceroute.cpp
@@ -120,13 +120,13 @@ class Tool : public nmdt::AbstractImportTool<P,R>
     {
       const auto& toolRunId {this->getToolRunId()};
       const auto& deviceId  {this->getDeviceId()};
-
+    
       for (auto& result : this->tResults) {
         // muck
 
         // save
         result.save(t, toolRunId, deviceId);
-        LOG_DEBUG << result.toDebugString() << std::endl;
+        LOG_DEBUG << result.toString() << std::endl;
 
         // link
       }
